Renamed Porthos::Caching to CachingShell
