+++ app/models/link_list.rb
@@ -1,5 +1,7 @@
 class LinkList
   include MongoMapper::Document
+  include CachingShell::Shelled
+
   key :title, String
   key :handle, String
   timestamps!
