--- lib/porthos/caching.rb
@@ -1,7 +1,0 @@
-module Porthos
-  module Caching
-    def self.shell_cache
-      @cache ||= ShellCache.instance
-    end
-  end
-end
