--- lib/porthos/caching/shell_cache.rb
@@ -1,27 +1,0 @@
-module Porthos
-  module Caching
-
-    class ShellCache
-      include Singleton
-
-      def set key, value
-        _cache[key] = value
-      end
-
-      def get key
-        _cache[key]
-      end
-
-      def clear
-        _cache.clear
-      end
-
-      private
-
-      def _cache
-        @_cache ||= {}
-      end
-    end
-
-  end
-end
